active archetype replacement

+++ go/cmd/orchestrator/manage_jobs.go
@@ -10,6 +10,7 @@ import (
 	"github.com/Jorrit05/DYNAMOS/pkg/etcd"
 	"github.com/Jorrit05/DYNAMOS/pkg/lib"
 	pb "github.com/Jorrit05/DYNAMOS/pkg/proto"
+	"github.com/google/uuid"
 )
 
 // /agents/jobs/SURF/jorrit.stutterheim@cloudnation.nl/jorrit-stutterheim-43ea82da
@@ -46,7 +47,6 @@ func deleteJobInfo(jobNames []string, userName string, changedAgreementName stri
 			}
 
 			compositionRequest := &pb.CompositionRequest{}
-
 			err = json.Unmarshal(resp.Kvs[0].Value, compositionRequest)
 			if err != nil {
 				logger.Sugar().Errorf("failed to unmarshal JSON: %v", err)
@@ -69,6 +69,7 @@ func checkJobs(agreement *api.Agreement) {
 
 		key := fmt.Sprintf("/agents/jobs/%s/%s", agreement.Name, relationName)
 
+		// Get all jobnames registered for this user of the data steward of this agreement
 		jobNames, err := etcd.GetKeysFromPrefix(etcdClient, key, etcd.WithMaxElapsedTime(2*time.Second))
 		if err != nil {
 			logger.Sugar().Warnf("error get agents: %v", err)
@@ -77,17 +78,19 @@ func checkJobs(agreement *api.Agreement) {
 			logger.Debug("no active jobs for this user")
 			return
 		}
+		for _, v := range jobNames {
+			logger.Sugar().Warnf(v)
+		}
 
-		if len(relationDetails.AllowedArchetypes) == 0 || relationDetails.AllowedArchetypes[0] == "" {
+		// New agreement has no allowed archetypes
+		if len(relationDetails.AllowedArchetypes) == 0 || (relationDetails.AllowedArchetypes[0] == "" && len(relationDetails.AllowedArchetypes) == 1) {
 			logger.Debug("This user no has no allowed archetypes")
 			if len(jobNames) > 0 {
 				deleteJobInfo(jobNames, relationName, agreement.Name)
 			}
-			// key := fmt.Sprintf("/agents/jobs/%s/queueInfo/%s", agreement.Name, activeJobCompositionRequests)
-
-			// activeJobCompositionRequests.de
 			continue
 		}
+		evaluateArchetypeInActiveJobs(jobNames, agreement, relationName, relationDetails, c)
 
 		// key := fmt.Sprintf("/agents/jobs/%s/%s/", agreement.Name, relationName)
 		// activeJobCompositionRequests, err := etcd.GetPrefixListEtcd(etcdClient, key, compositionRequest)
@@ -95,22 +98,213 @@ func checkJobs(agreement *api.Agreement) {
 		// 	logger.Sugar().Warnf("error get jobs: %v", err)
 		// }
 
-		// evaluateArchetypeChoice(activeJobCompositionRequests, agreement, relationDetails)
 	}
 }
 
-func evaluateArchetypeChoice(activeJobCompositionRequests []*pb.CompositionRequest, agreement *api.Agreement, relationDetails api.Relation) {
-	logger.Debug("starting evaluateArchetypeChoice")
+func evaluateArchetypeInActiveJobs(jobNames []string, agreement *api.Agreement, relationName string, relationDetails api.Relation, c pb.SideCarClient) {
+	logger.Debug("starting evaluateArchetypeInActiveJobs")
+	ctx := context.Background()
 	// alue.ArchetypeId == archetype in current active job from the agreement name.
 
-	for _, value := range activeJobCompositionRequests {
-		if lib.NewSet(relationDetails.AllowedArchetypes).Has(value.ArchetypeId) {
-			println("still matching archetypes")
+	// for each job. Check current archetype. versus new archetypes.
+	for _, job := range jobNames {
+
+		jobInfoKey := fmt.Sprintf("/agents/jobs/%s/%s/%s", agreement.Name, relationName, job)
+
+		resp, err := etcdClient.Get(ctx, jobInfoKey)
+		if err != nil {
+			logger.Sugar().Errorf("error getting value from etcd: %v", err)
+			continue
+		}
+
+		if len(resp.Kvs) == 0 {
+			logger.Warn("this should not happen")
+			continue
+		}
+
+		currentRegisteredJob := &pb.CompositionRequest{}
+		err = json.Unmarshal(resp.Kvs[0].Value, currentRegisteredJob)
+		if err != nil {
+			logger.Sugar().Errorf("error unmarshalling jobinfo: %v", err)
+		}
+
+		policyUpdate := &pb.PolicyUpdate{
+			Type:            "policyUpdate",
+			User:            &pb.User{Id: relationDetails.ID, UserName: relationName},
+			RequestMetadata: &pb.RequestMetadata{DestinationQueue: "policyEnforcer-in"},
+		}
+
+		correlationId := uuid.New().String()
+		policyUpdate.RequestMetadata.CorrelationId = correlationId
+
+		agentsWithThisJob := make(map[string]*pb.CompositionRequest)
+
+		ctx = getJobAcrossAgents(ctx, agentsWithThisJob, job, relationName)
+
+		for k, v := range agentsWithThisJob {
+			if v.Role == "all" || v.Role == "dataProvider" {
+				policyUpdate.DataProviders = append(policyUpdate.DataProviders, k)
+			}
+		}
+
+		policyUpdateMutex.Lock()
+		policyUpdateMap[policyUpdate.RequestMetadata.CorrelationId] = agentsWithThisJob
+		policyUpdateMutex.Unlock()
+		c.SendPolicyUpdate(ctx, policyUpdate)
+	}
+}
+
+func processPolicyUpdate(ctx context.Context, agentsWithThisJob map[string]*pb.CompositionRequest, policyUpdate *pb.PolicyUpdate) {
+	logger.Sugar().Debugf("processPolicyUpdate")
+
+	archetype, err := chooseArchetype(policyUpdate.ValidationResponse.ValidDataproviders)
+	if err != nil {
+		logger.Sugar().Errorf("error choosing archetype: %v", err)
+	}
+
+	logger.Sugar().Debugf("New archetype: %v", archetype)
+
+	var archetypeConfig api.Archetype
+	_, err = etcd.GetAndUnmarshalJSON(etcdClient, fmt.Sprintf("/archetypes/%s", archetype), &archetypeConfig)
+	if err != nil {
+		logger.Sugar().Errorf("error choosing archetype: %v", err)
+		return
+	}
+
+	// technically now, this shouldn't be necessary
+	computeProviderAlready := false
+	var ttp lib.AgentDetails
+	for agent, currentData := range agentsWithThisJob {
+		if currentData.ArchetypeId == archetype {
+			logger.Sugar().Debug("same archetype, do nothing")
+			return
+		}
+		key := fmt.Sprintf("/agents/jobs/%s/%s/%s", agent, policyUpdate.User.UserName, currentData.JobName)
+
+		if archetypeConfig.ComputeProvider != "other" {
+			if currentData.Role == "computeProvider" {
+				// Delete this job info
+				_, err := etcdClient.Delete(ctx, key)
+				if err != nil {
+					logger.Sugar().Warnf("error deleting key from etcd: %v", err)
+				}
+				continue
+			}
+
+			// New archetype is computeToData
+			newData := currentData
+			newData.ArchetypeId = archetype
+			newData.Role = "all"
+			newData.DataProviders = []string{}
+			err := etcd.SaveStructToEtcd[*pb.CompositionRequest](etcdClient, key, newData)
+			if err != nil {
+				logger.Sugar().Errorf("Error saving struct to etcd: %v", err)
+				return
+			}
+			computeProviderAlready = true
 		} else {
-			println("previous archetype no longer allowed")
-			// chooseArchetype()
-			// Now  I think I can use existing functions to determine a new  archetype
+			var err error
+			ttp, err = chooseThirdParty(policyUpdate.ValidationResponse)
+			if err != nil {
+				logger.Sugar().Errorf("Error choosing third party: %v", err)
+				return
+			}
+
+			if currentData.Role == "computeProvider" && agent == ttp.Name {
+				computeProviderAlready = true
+				continue
+			} else if currentData.Role == "computeProvider" && agent != ttp.Name {
+				// Delete this job info
+				_, err := etcdClient.Delete(ctx, key)
+				if err != nil {
+					logger.Sugar().Warnf("error deleting key from etcd: %v", err)
+				}
+				continue
+			}
+
+			if currentData.Role == "all" {
+				_, ok := policyUpdate.ValidationResponse.ValidDataproviders[agent]
+				if !ok {
+					// Delete this job info
+					_, err := etcdClient.Delete(ctx, key)
+					if err != nil {
+						logger.Sugar().Warnf("error deleting key from etcd: %v", err)
+					}
+				}
+
+				// New archetype is dataThroughTtp
+				newData := currentData
+				newData.ArchetypeId = archetype
+				newData.Role = "dataProvider"
+				newData.DataProviders = []string{}
+
+				err = etcd.SaveStructToEtcd[*pb.CompositionRequest](etcdClient, key, newData)
+				if err != nil {
+					logger.Sugar().Errorf("Error saving struct to etcd: %v", err)
+					return
+				}
+
+			}
+
 		}
 	}
 
+	if !computeProviderAlready {
+		compositionRequest := &pb.CompositionRequest{}
+		compositionRequest.User = policyUpdate.User
+		tmpDataProvider := []string{}
+
+		for key := range policyUpdate.ValidationResponse.ValidDataproviders {
+			tmpDataProvider = append(tmpDataProvider, key)
+		}
+		compositionRequest.Role = "computeProvider"
+		compositionRequest.DataProviders = tmpDataProvider
+		compositionRequest.ArchetypeId = archetype
+		for _, v := range agentsWithThisJob {
+			compositionRequest.RequestType = v.RequestType
+			compositionRequest.JobName = v.JobName
+			break
+		}
+
+		compositionRequest.DestinationQueue = ttp.RoutingKey
+
+		c.SendCompositionRequest(ctx, compositionRequest)
+	}
+}
+
+func getJobAcrossAgents(ctx context.Context, targetMap map[string]*pb.CompositionRequest, jobName string, userName string) context.Context {
+
+	var agents *lib.AgentDetails
+	key := "/agents/online/"
+	activeAgents, err := etcd.GetPrefixListEtcd(etcdClient, key, agents)
+	if err != nil {
+		logger.Sugar().Warnf("error get agents: %v", err)
+	}
+
+	for _, agent := range activeAgents {
+
+		key := fmt.Sprintf("/agents/jobs/%s/%s/%s", agent.Name, userName, jobName)
+
+		resp, err := etcdClient.Get(ctx, key)
+		if err != nil {
+			logger.Sugar().Errorf("error getting value from etcd: %v", err)
+			continue
+		}
+
+		if len(resp.Kvs) == 0 {
+			logger.Sugar().Debugw("no value found for", "key", key)
+			continue
+		}
+
+		agentsConfiguration := &pb.CompositionRequest{}
+		err = json.Unmarshal(resp.Kvs[0].Value, agentsConfiguration)
+		if err != nil {
+			logger.Sugar().Errorf("error unmarshalling jobinfo: %v", err)
+			continue
+		}
+
+		targetMap[agent.Name] = agentsConfiguration
+	}
+
+	return ctx
 }
